Re #25: Fix GUI button display issue
The bias, continuity, and tension inputs remained visible if closed
surface mode was selected immediately after selecting the Kochanek
spline interpolation option in curve mode.

This is because the radio button for Kochanek spline interpolation is
still selected even though it is no longer visible. This change modifies
the conditionals so that these inputs are only shown if both Kochanek
spline interpolation *and* curve mode are selected.