Draw model even when surface is open
When the surface is open (the points are on the same plane), previously
no model could be drawn. This change detects open surfaces and creates
new points in a cube shape surrounding the ones placed by the user,
so that a (mostly flat) 3-dimensional model can be drawn.

Additional points are not added if the surface is already closed.
Once there are enough points that the surface becomes closed, the check
will not be repeated upon placement of additional points.

This change means that a minimum number of fiducials is no longer
necessary for drawing a model surface - even if there are between zero
and four fiducial points, a plane is formed and a model can be drawn.