Added option to compute warping transform to parent
When models are transformed, "ToParent" transform is used, therefore transforming large models can be made faster if warping transform is "ToParent" is computed directly (and not computed from its inverse).

When images are transformed, "FromParent" transform is used (as image transformation requires resampling). Therefore, for transforming images it is better to keep the WarpingTransformFromParent flag enabled (that's the default value).

The option is currently not exposed on the GUI, as it is an advanced setting. If it proves to be important to allow users to change this setting without using the Python console then it can be added to the module GUI.