ENH: Make guidelet layouts dynamic
Instead of registering the layouts in the Guidelet base class in a static way, layouts can now be registered both from the Guidelet class and the application subclass using registerLayout. It takes four arguments: layoutName (as it appears in the combo box), layoutId (integer), the layout XML, and an optional callback function that is called when the layout is set (for example fit ultrasound in red slice and show it in 3D).