BUG: Rename cleanup function so that it happens at the desired time
Since the commit https://github.com/Slicer/Slicer/commit/33f09a72b4fac7d13e793a83e769ccca6f246693 in Slicer core, the scripted loadable module base class has a cleanup function that is called just before unloading the module. Unfortunately that is too late, because by then most of the Qt objects are destructed, so attempts at unsetting and disconnecting fail with an error message. In order for it to succeed, it needs to have a different name.

Note that this change will need to be propagated to all the existing guidelet subclasses.