BUG: Fix cleanup issues when exiting guidelets also in GuideletLoadable
This is the last piece of the work that fixes the errors on exiting a guidelet. The cleanup function needed to be renamed here as well so that it is not called at the wrong time (when unloading the module from memory which is too late), and instead it is done at hideEvent.