Fix resetting node selectors in Breach Warning module
While tracking was active and input transform and parent transform of the input model was being constantly modified, the node selectors kept switching back to the current selection (see https://discourse.slicer.org/t/issue-with-breachwarning-in-slicer-4-11/17456/17). An extra check prevents these continuous updates and only updates the node selector if selection in the combobox is finalized.