BUG: Fix ROI center when calling CalculateROIFromVolumeSequence
When calling CalculateROIFromVolumeSequence with a Markups ROI node that has a center that is not [0,0,0], the center of the ROI would be incorrect.
This was because for Markups ROI the bounds were calculated in the "Object" coordinate system, while the SetCenter method expects coordinates in "Node" coordinates.

Fixed by transforming the center from Object to Node before calling SetCenter.