ENH: Ignore invalid transforms from OpenIGTLink during pivot calibration
If tracking is lost during pivot calibration and Plus is configured to send invalid transforms, it will cause the pivot calibration to fail as the transform will jump to the identity matrix.
Now we also check the "OpenIGTLink.TransformValid" and skip adding the transform to the pivot calibration if it is not valid.