BUG: Use vtkMRMLTransformNode instead of vtkMRMLLinearTransformNode
When the user adds a new transform in Transforms module, then now a vtkMRMLTransformNode node is created (instead of vtkMRMLLinearTransformNode).
vtkMRMLTransformNode is preferable because a transform node must be able to store arbitrary VTK transforms in it (because transforms can be hardened on each other and transforms can be split).

However, GUI and API of several SlicerIGT module only supported vtkMRMLLinearTransformNode. This resulted in transforms not being selectable in various SlicerIGT modules (see for example https://discourse.slicer.org/t/linear-transform-not-appearing-anymore-as-a-driver-in-igt-volume-reslicer-driver/45497).

Fixed by using vtkMRMLTransformNode instead of vtkMRMLLinearTransformNode everywhere.